refactor: fix fullscreen provider
